feat(test): add e2e tests with consensus and execution layers (#433)
This patch adds a first set of tests to launch tempo nodes with a
consensus and execution layer.

The tests are:

1. 5 nodes with perfect linkage; test success reaching height 5
2. 5 nodes with 75% good links (25% failure); test success is reaching
height 5
3. 10 nodes with 98% good links; test success is reaching height 20

The tests were written very close to commonware's examples:

1. alto:
https://github.com/commonwarexyz/alto/blob/a831c6eab7ec10451dca025cedb34d25ccb0c2f6/chain/src/lib.rs#L42
2. monorepo:
https://github.com/commonwarexyz/monorepo/blob/d049d43c3efe3a6a18e379115e587bb9e03bf895/consensus/src/threshold_simplex/mod.rs#L210

## Downsides of this test setup

1. runs are not completely deterministic. Ideally, we two successive
runs with the same input parameters return the same audit result (a hash
of all metrics collected by the runtime). However, as we are connecting
a deterministic (commonware) runtime to a non-deterministic (tokio/reth)
runtime, the completion of futures is not always guaranteed, even if
reth always returns the same output given the same inputs. The source of
this non-determinism is difficult to pinpoint. To support this feature,
commonware recently merged the new `external` feature [1] that allows
"pacing" certain futures, making the deterministic runtime await them in
real time (and not in simulated time, as it normally would).
2. opening too many reth nodes leads to errors like `failed to open the
database: unknown error code: 12 (12)`. This only happens if very many
runs are launched righ after another, as is the case in the commonware
examples, where runs are repeated in a loop with different seeds.
Running the tests in nextest seems to help.
3. Runs take a long time. This is partially because of waiting for real
time. Given a proposal time of 2s and a verification time of 1s, one can
expect 1 block to take up to 3s in real time. So to reach 5 blocks twice
(to verifiy audits) can take of the order of 30s. If one now also
includes simulated jitter and communication failures
(expected/preconfigured) as well as non-determinism issues (unexpected),
this number is expected to go up. Reaching 20 blocks (as is the case for
one test) should take about 40s, which indeed it does.

## Note on seemingly orthgonal changes in this PR

This patch adds a non-zero `gasLimit` entry to the `ADAGIO` hardcoded
genesis/chainspec. `gasLimit` being absent meant that it took a default
value of 0, which made it impossible to run tests. The likely failure
point is this check:
:https://github.com/tempoxyz/tempo/blob/7166a3d8cfbab6a5174e706a079cfd4f1bee2ae1/crates/consensus/src/lib.rs#L57


[1]: https://github.com/commonwarexyz/monorepo/pull/1848

---------

Co-authored-by: Patrick O'Grady <me@patrickogrady.xyz>