chore: remove compact hack for mainnet (#1814)
We introduced KeychainAuthorization in the tempo transaction in a
hardfork. To do this in a backward compatible way, we added a non
standard compact encoding for Key Auths with a note to remove before
mainnet.

This PR removes the custom compact implementation, and reverts to the
standard compact macro.

Addresses #1429